# api/openapi.yaml
openapi: 3.1.0
info:
  title: Signing Platform API
  version: 0.1.0
  description: REST API for the original signing platform.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/me:
    get:
      operationId: getCurrentUser
      summary: Synchronize and return the authenticated application user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authenticated local user and active WorkOS organization context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Missing or invalid WorkOS access token
        "502":
          description: The WorkOS user directory could not be reached
  /v1/envelopes:
    get:
      operationId: listEnvelopes
      summary: List envelopes in the active workspace
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, in_progress, completed, voided, expired]
      responses:
        "200": { description: Workspace envelopes }
        "401": { description: Missing or invalid access token }
    post:
      operationId: createEnvelope
      summary: Create a draft envelope in the active workspace
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                language: { type: string, default: en }
                timezone: { type: string, default: Etc/UTC }
                dateFormat: { type: string, default: YYYY-MM-DD }
                expiresAt: { type: string, format: date-time }
      responses:
        "201": { description: Draft envelope created }
        "400": { description: Invalid request }
        "401": { description: Missing or invalid access token }
  /v1/envelopes/{envelopeId}:
    get:
      operationId: getEnvelope
      summary: Get an envelope from the active workspace
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: envelopeId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Envelope }
        "401": { description: Missing or invalid access token }
        "404": { description: Envelope does not exist in this workspace }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CurrentUser:
      type: object
      required: [user]
      properties:
        user:
          type: object
          required: [id, email, name, provider]
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            name:
              type: string
            emailVerifiedAt:
              type: string
              format: date-time
            provider:
              type: string
              example: workos
        workspace:
          type: object
          required: [id, name, slug, role]
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            slug:
              type: string
            role:
              type: string
        permissions:
          type: array
          items:
            type: string